refactor: 쓰기 읽기 책임 분리 및 정산 관련 Factory 추상화
